multimodal-search-pdf/requirements.txt
click==7.1.2
transformers==4.1.1
torch==1.7.1
jina[scipy, http]==1.0.0
PyPDF2==1.26.0
PyMuPDF==1.18.5
Pillow==8.0.1
